requirements.txt
transformers>=4.30.0
torch>=2.0.0
tensorflow>=2.12.0
flask>=2.0.0
requests>=2.28.0
numpy>=1.24.0
pillow>=9.5.0
soundfile>=0.12.1
librosa>=0.10.0
opencv-python>=4.7.0
python-dotenv>=1.0.0
groq>=0.3.0
langchain>=0.1.0
langchain-groq>=0.0.1 